continuando storia 3

+++ database/migrations/2023_09_05_190918_create_announcements_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('announcements', function (Blueprint $table) {
             $table->id();
             $table->string('title');
-            $table->float('price',8,2);
+            $table->decimal('price',8,2);
             $table->longText('description');
             $table->foreignId('user_id')->nullable()->constrained()->onDelete('cascade');
